pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vaanidub"
version = "0.1.0"
description = "AI-powered regional dubbing tool for Indian languages"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{name = "VaaniDub Team"}]

dependencies = [
    # Core
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "structlog>=23.0",

    # Audio/Video processing
    "ffmpeg-python>=0.2.0",
    "pydub>=0.25.1",
    "soundfile>=0.12.1",
    "scipy>=1.11",
    "numpy>=1.24",

    # API
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "python-multipart>=0.0.9",

    # Database
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "aiosqlite>=0.19",

    # CLI
    "typer[all]>=0.9",
    "rich>=13.0",

    # Utilities
    "httpx>=0.27",
    "tenacity>=8.2",
]

[project.optional-dependencies]
gpu = [
    # STT + Diarization
    "faster-whisper>=1.0.0",
    "whisperx @ git+https://github.com/m-bain/whisperX.git",
    "pyannote.audio>=3.1",
    # Audio analysis
    "librosa>=0.10.0",
    "praat-parselmouth>=0.4.3",
    "pyrubberband>=0.3.0",
    # Source separation
    "demucs>=4.0",
    # Translation / TTS (HuggingFace models)
    "transformers>=4.40",
    "accelerate>=0.30",
    "sentencepiece>=0.1.99",
    "protobuf>=4.0",
    # PyTorch
    "torch>=2.2",
    "torchaudio>=2.2",
]
worker = [
    "celery[redis]>=5.3",
    "redis>=5.0",
]
postgres = [
    "psycopg2-binary>=2.9",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.3",
    "mypy>=1.8",
]
elevenlabs = [
    "elevenlabs>=1.0",
]
google = [
    "google-cloud-translate>=3.15",
]
all = [
    "vaanidub[gpu,worker,postgres,elevenlabs,google]",
]

[project.scripts]
vaanidub = "vaanidub.cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["vaanidub*"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.10"
strict = true
